TAT-868: using selenium grid

+++ envs/common/steps/login.js
@@ -9,25 +9,30 @@ if (steps.login === undefined) {
                 console.log('User: ' + _tf_config.user_name);
                 pages.login.open(true);
 
-                browser.manage().getCookies().then((cookies) => {
-                    if (!_.find(cookies, {name: 'TAT_Login'})) {
-                        pages.login.login();
+                if (systemConfig.persistProfile) {
+                    browser.manage().getCookies().then((cookies) => {
+                        if (!_.find(cookies, {name: 'TAT_Login'})) {
+                            pages.login.login();
 
-                        pages.login.waitForAjax();
-                        pages.login.check();
+                            pages.login.waitForAjax();
+                            pages.login.check();
 
-                        browser.manage().getCookies().then((cookies) => {
-                            let cookie = _.find(cookies, {name: 'JSESSIONID'});
+                            browser.manage().getCookies().then((cookies) => {
+                                let cookie = _.find(cookies, {name: 'JSESSIONID'});
 
-                            if (cookie) {
-                                browser.manage().deleteCookie(cookie.name);
-                                browser.manage().addCookie(cookie.name, cookie.value, undefined, undefined, false, futureDate);
-                                browser.manage().addCookie('TAT_Login', true, undefined, undefined, false, futureDate);
-                            }
-                        });
-                    }
-                });
-
+                                if (cookie) {
+                                    browser.manage().deleteCookie(cookie.name);
+                                    browser.manage().addCookie(cookie.name, cookie.value, undefined, undefined, false, futureDate);
+                                    browser.manage().addCookie('TAT_Login', true, undefined, undefined, false, futureDate);
+                                }
+                            });
+                        }
+                    });
+                } else {
+                    pages.login.login();
+                    pages.login.waitForAjax();
+                    pages.login.check();
+                }
 
             });
         },
